requirements.txt
Flask==2.0.1
requests==2.26.0
python-dotenv==0.18.0
black==21.7b0
python-dateutil==2.8.2
diskcache==5.2.1
tqdm==4.62.0
Jinja2==3.0.1
gunicorn==20.1.0
Werkzeug==2.2.2
